edit user notification api
