+++ Back-end/Capstone.Model/UserNotification.cs
@@ -18,6 +18,7 @@ namespace Capstone.Model
         public User User { get; set; }
 
         public bool IsRead { get; set; }
+        public bool IsHandled { get; set; }
         public bool IsDeleted { get; set; }
     }
 }
